requirements.txt
PyQt6
cryptography
bcrypt
python-dotenv